chat cli wip #pr
